ec/delete-my-account: check functions are not coroutines
If the check is a coroutine, then the check effectively becomes lambda message: True
since all coroutine objects are truthy. This command still worked on my machine, though,
because the only messages that the test bot could see were messages in the testing server pretty much.

So as soon as the bot saw a new message, it would assume confirmation.