extensions/db: fix all_emotes_keyset(before=)
So previously I was getting the first 100 emotes before the provided name, *sorted alphabetically*.
Well that's just the first 100 emotes no matter what. It doesn't matter if you say "first 100 before #101"
or "first 100 before #2000", that's still the first 100.

So to fix that, we do it in reverse, and then reverse it client side. (Is this still as efficient?)