fix utils.gather_or_cancel
It literally did not do what it says it would do at all.
For some reason the old technique did not work, contradicting the python documentation.

The new technique is garbo and it makes me want to use Trio instead where gather_or_cancel is the default behavior
of nurseries.